video-to-notes-generator/requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.8
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@e58f28804528831904c3b6f2c0e473f346223433
pillow==10.2.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.3
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
sympy==1.13.2
tbb==2021.13.1
tiktoken==0.7.0
torch==2.4.0
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
